Fix broken firing animation
The problem with the animation was that it appeared to be firing multiple shots when you only shot one. This has been fixed by using a Trigger animation parameter and turning off looping for the firing animation.

Note that this is not a perfect solution and a UnityEvent on the Weapon class may be a better option.