PlayerMovement.cs, PlayerRotation.cs - More appropriate naming + controller rotation support
This commit renames PlayerMouseLook to PlayerRotation to reflect its new controller support as well as PlayerController to PlayerMovement to be more consistent with the new naming. This commit also adds the ability to support both Keyboard/Mouse and controller for rotation.

This commit can probably be cleaned and simplified a little however this commit just includes the basic implementation.